heroku rodrigoapi/requirements.txt
Django
djangorestframework
django-cors-headers
gunicorn
psycopg2
whitenoise